Start tracking user active times and deleting study sessions that ...
were left unfinished (like when the user closes
their browser while the timer is still running).

The timer is on the client side, so when the student
closes their browser, there really is no way to
communicate that the study session is over.
And so it never resumes until the student becomes
active again.

That's why I'm querying for such 'forgotten' sessions and
deleting them. I would've otherwise done this via
a cron job but heroku wants me to pay for that, so
I've decided to just run that query on any AJAX update
for a user's activity. It's stupid, I know.